+++ benchmark/scripts/benchmark_dpo_loss.py
@@ -1,5 +1,6 @@
+from test.chunked_loss.test_dpo_loss import HF_DPO_Loss
+
 import torch
-import triton
 from utils import (
     QUANTILES,
     SingleBenchmarkRunInput,
@@ -9,7 +10,8 @@ from utils import (
     run_benchmarks,
 )
 
-from liger_kernel.alignment.dpo_loss import HF_DPO_Loss, LigerFusedLinearDPOFunction
+import triton
+from liger_kernel.chunked_loss.dpo_loss import LigerFusedLinearDPOFunction
 
 
 class TorchDPOLoss(torch.nn.Module):
